feat(fe): delete pagination and separate SearchBar as component
